- Convert deprecated xproj project format to csproj - Add .NET Standard 2.0 as a project target - Add generic Serilog Sink icon - Remove empty license file - Remove GitCheese reference - Minor reformatting - Move model and option files to Models directory
