--- Serilog.Sinks.Slack.Sample/project.json
@@ -1,22 +1,0 @@
-{
-  "version": "1.0.0-*",
-  "buildOptions": {
-    "emitEntryPoint": true
-  },
-  "dependencies": {
-    "Microsoft.NETCore.App": {
-      "type": "platform",
-      "version": "1.1.1"
-    },
-    "Serilog.Sinks.Console": "2.1.0",
-    "Serilog.Sinks.Slack": {
-      "target": "project"
-    }
-  },
-
-  "frameworks": {
-    "netcoreapp1.0": {
-      "imports": "dnxcore50"
-    }
-  }
-}
