+++ Serilog.Sinks.Slack.sln
@@ -1,19 +1,17 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30413.136
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{626813A1-D71D-43F8-ABCC-D3CBF0189A3E}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{721763D1-0FA7-42C4-A71B-91BBE31BB57D}"
 	ProjectSection(SolutionItems) = preProject
 		appveyor.yml = appveyor.yml
-		global.json = global.json
-		LICENSE.md = LICENSE.md
 		README.md = README.md
 	EndProjectSection
 EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Serilog.Sinks.Slack", "Serilog.Sinks.Slack\Serilog.Sinks.Slack.xproj", "{518E39AE-AD6E-44DC-BECE-7C5F68A88C95}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Serilog.Sinks.Slack", "Serilog.Sinks.Slack\Serilog.Sinks.Slack.csproj", "{518E39AE-AD6E-43DC-BECE-7C5F68A88C95}"
 EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Serilog.Sinks.Slack.Sample", "Serilog.Sinks.Slack.Sample\Serilog.Sinks.Slack.Sample.xproj", "{E2BDF0A2-DAE4-4930-84ED-A095FD7F349C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Serilog.Sinks.Slack.Sample", "Serilog.Sinks.Slack.Sample\Serilog.Sinks.Slack.Sample.csproj", "{E2BDF0A2-DAE4-2930-84ED-A095FD7F349C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,14 +19,14 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{518E39AE-AD6E-44DC-BECE-7C5F68A88C95}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{518E39AE-AD6E-44DC-BECE-7C5F68A88C95}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{518E39AE-AD6E-44DC-BECE-7C5F68A88C95}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{518E39AE-AD6E-44DC-BECE-7C5F68A88C95}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E2BDF0A2-DAE4-4930-84ED-A095FD7F349C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E2BDF0A2-DAE4-4930-84ED-A095FD7F349C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E2BDF0A2-DAE4-4930-84ED-A095FD7F349C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E2BDF0A2-DAE4-4930-84ED-A095FD7F349C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{518E39AE-AD6E-43DC-BECE-7C5F68A88C95}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{518E39AE-AD6E-43DC-BECE-7C5F68A88C95}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{518E39AE-AD6E-43DC-BECE-7C5F68A88C95}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{518E39AE-AD6E-43DC-BECE-7C5F68A88C95}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E2BDF0A2-DAE4-2930-84ED-A095FD7F349C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E2BDF0A2-DAE4-2930-84ED-A095FD7F349C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E2BDF0A2-DAE4-2930-84ED-A095FD7F349C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E2BDF0A2-DAE4-2930-84ED-A095FD7F349C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
