+++ Serilog.Sinks.Slack/Models/Attachment.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using Newtonsoft.Json;
 
-namespace Serilog.Sinks.Slack
+namespace Serilog.Sinks.Slack.Models
 {
     public class Attachment
     {
