--- Serilog.Sinks.Slack/project.json
@@ -1,49 +1,0 @@
-{
-  "version": "1.2.0",
-  "title": "Serilog.Sinks.Slack",
-  "dependencies": {
-    "Newtonsoft.Json": "9.0.1",
-    "Serilog": "2.4.0",
-    "Serilog.Sinks.PeriodicBatching": "2.1.0"
-  },
-  "authors": [ "mgibas", "Sergey Kuznetsov" ],
-  "copyright": "Copyright 2017",
-  "packOptions": {
-    "summary": "Serilog sink for Slack",
-    "tags": [ "Serilog", "Sink", "Slack" ],
-    "owners": [ "gibasmaciej", "phnx47" ],
-    "projectUrl": "https://github.com/mgibas/serilog-sinks-slack",
-    "licenseUrl": "https://opensource.org/licenses/MIT",
-    "repository": {
-      "type": "git",
-      "url": "https://github.com/mgibas/serilog-sinks-slack"
-    }
-  },
-  "frameworks": {
-    "net45": {
-      "frameworkAssemblies": {
-        "System.Net.Http": "4.0.0.0"
-      }
-    },
-    "net46": {
-      "frameworkAssemblies": {
-        "System.Net.Http": "4.0.0.0"
-      }
-    },
-    "netstandard1.1": {
-      "dependencies": {
-        "System.Net.Http": "4.3.0",
-        "System.Runtime.InteropServices": "4.3.0"
-      }
-    },
-    "netstandard1.3": {
-      "dependencies": {
-        "System.Net.Http": "4.3.0",
-        "System.Runtime.InteropServices": "4.3.0"
-      }
-    }
-  },
-  "tools": {
-    "dotnet-version": "1.1.0"
-  }
-}
