--- global.json
@@ -1,9 +1,0 @@
-{
-  "sdk": {
-    "version": "1.0.0-preview2-1-003177"
-  },
-  "projects": [
-    "Serilog.Sinks.Slack",
-    "Serilog.Sinks.Slack.Sample"
-  ]
-}
